# docs/swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "schemas": {
      "ApiErrors": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "FornecedorRequest": {
        "type": "object",
        "required": ["razaoSocial", "cnpj", "email", "telefone", "situacao"],
        "properties": {
          "razaoSocial": {
            "type": "string",
            "pattern": "\\S"
          },
          "cnpj": {
            "type": "string",
            "pattern": "\\S"
          },
          "email": {
            "type": "string",
            "pattern": "\\S"
          },
          "telefone": {
            "type": "string",
            "pattern": "\\S"
          },
          "situacao": {
            "$ref": "#/components/schemas/SituacaoFornecedorEnum"
          }
        }
      },
      "FornecedorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "razaoSocial": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "situacao": {
            "$ref": "#/components/schemas/SituacaoFornecedorEnum"
          },
          "dataBaixa": {
            "$ref": "#/components/schemas/LocalDate"
          }
        }
      },
...
      "LocalDate": {
        "type": "string",
        "format": "date",
        "examples": ["2022-03-10"]
      },
...
      "SituacaoFornecedorEnum": {
        "type": "string",
        "enum": ["ATIVO", "BAIXADO", "SUSPENSO"]
      },
...
    }
  },
  "paths": {
    "/api/fornecedor": {
      "get": {
        "parameters": [
          {
            "name": "cnpj",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "razaoSocial",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "situacao",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de fornecedores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FornecedorResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Get All",
        "tags": ["Fornecedor Controller"]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FornecedorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Fornecedor criado com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Create",
        "tags": ["Fornecedor Controller"]
      }
    },
    "/api/fornecedor/{id}": {
      "put": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FornecedorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Fornecedor atualizado com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Update",
        "tags": ["Fornecedor Controller"]
      },
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de fornecedores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FornecedorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Get By Id",
        "tags": ["Fornecedor Controller"]
      },
      "delete": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fornecedor excluído com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Delete",
        "tags": ["Fornecedor Controller"]
      }
    },
    "/api/fornecedor/{id}/baixa": {
      "patch": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fornecedor baixado com sucesso"
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrors"
                }
              }
            }
          }
        },
        "summary": "Baixar",
        "tags": ["Fornecedor Controller"]
      }
    },
...
  },
  "info": {
    "title": "Fisctrack API",
    "version": "1.0",
    "description": "API para gerenciamento de notas fiscais",
    "contact": {
      "name": "Suporte",
      "email": "wiriswernek@gmail.com",
      "url": "https://wiriswernek.tech"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}